refactor(Knowledge): 梳理清减 Knowledge 模块，消除重复、统一配置类、正交分解 API 层;
- 消除 5 处重复代码：将 _infer_source_type 和 _normalize_source_metadata 合并至 types.py 单一事实源
- 删除死代码：移除无引用的 GraphSearchMatch 类型和 summarization.py 兼容层
- 统一配置类：GraphBuildConfigModel → GraphBuildConfig、GraphSearchConfig → GraphQueryConfig，消除 dataclass/Pydantic 二元体系
- 正交分解：提取 api.py 中 42 个 Pydantic 模型至 schemas.py，实现路由与模型正交
- 保留 __init__.py 向后兼容别名（GraphBuildConfigModel、GraphSearchConfig）
- 净减约 300 行源码（-531 / +237）
